Fix Improve UX to recover files (fix #4130)
Before this solution, every time a session was saved, the process ID
('pid' number) was also saved in a file called 'pid'. During Aseprite
startup, the 'pid' file of each session was read, each number was taken
to check if there was a process with that 'pid' number running on
the operating system.
The problem was that 'is_running_process()' did not discriminate
whether the 'pid' to be checked corresponded to Aseprite or not,
it was only checked if it was running. So if a 'pid' number of a
candidate session to be recovered was saved with a 'pid' number
matching any running process, Aseprite would hide such session.